requirements.txt
torch>=2.0
torchvision>=0.15
torchaudio>=2.0
timm>=1.0.0
scikit-learn>=1.1
scikit-image>=0.20
opencv-python>=4.5
albumentations>=1.3
einops
matplotlib
numpy
pandas
tqdm
pyyaml
tensorboard
rasterio
fiona
shapely
pyproj
rtree
